refactor: #21 - 출력 메서드 리팩터링 기존의 출력 담당 메서드와 테이블 형태로 출력하는 클래스의 역할이 중복되기 때문에 테이블 출력 클래스를 View의 inner 클래스로 변경
